@@ -11,11 +11,10 @@
   "dependencies": {
     "@editorjs/code": "^2.7.0",
     "@editorjs/editorjs": "2.19.3",
+    "@editorjs/image": "^2.6.2",
     "@emotion/react": "^11.7.1",
     "@emotion/styled": "^11.6.0",
     "@hookform/resolvers": "^2.8.8",
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
     "@mui/icons-material": "^5.4.1",
     "@mui/lab": "^5.0.0-alpha.70",
     "@mui/material": "^5.4.1",
@@ -25,7 +24,7 @@
     "babel-preset-next": "^1.4.0",
     "clsx": "^1.1.1",
     "macro-css": "^1.0.5",
-    "next": "^11.1.3",
+    "next": "^12.1.0",
     "next-redux-wrapper": "^7.0.5",
     "nookies": "^2.5.2",
     "react": "17.0.2",
@@ -36,7 +35,7 @@
     "yup": "^0.32.11"
   },
   "devDependencies": {
-    "@editorjs/image": "^2.6.2",
+    "@types/node": "^17.0.24",
     "eslint": "7.32.0",
     "eslint-config-next": "11.1.0",
     "typescript": "^4.3.5"
